requirements.txt
# Data collection and processing
pandas>=2.0
numpy>=1.24
soccerdata>=1.0
requests>=2.28
beautifulsoup4>=4.12
lxml>=4.9

# Database
sqlalchemy>=2.0

# Modeling
scipy>=1.10
scikit-learn>=1.3
lightgbm>=4.0

# Visualization
matplotlib>=3.7
seaborn>=0.12

# Jupyter
jupyterlab>=4.0
ipykernel>=6.0

# Configuration
pyyaml>=6.0

# Testing
pytest>=7.0
